pyproject.toml
[tool.poetry]
name = "users-groups"
version = "0.1.0"
description = ""
authors = ["Venu Vardhan Reddy Tekula <venuvardhanreddytekula8@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.5"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
django-phonenumber-field = "^6.0.0"
phonenumbers = "^8.12.42"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
makemigrations = "scripts:makemigrations"
migrate = "scripts:migrate"
server = "scripts:server"
shell = "scripts:shell"
tests = "scripts:tests"
createsuperuser = "scripts:createsuperuser"
